@@ -2,15 +2,16 @@ import { Injectable, Inject, Optional } from '@angular/core';
 import { OAuthService } from '../oauth-service';
 import { OAuthStorage } from '../types';
 import { HttpEvent, HttpHandler, HttpInterceptor, HttpRequest, HttpResponse, HttpErrorResponse } from '@angular/common/http';
-import {Observable} from 'rxjs/Observable';
-import { map, catchError } from 'rxjs/operators';
-import { of } from 'rxjs/observable/of';
+import { Observable } from 'rxjs';
+import { catchError } from 'rxjs/operators';
 import { OAuthResourceServerErrorHandler } from "./resource-server-error-handler";
 import { OAuthModuleConfig } from "../oauth-module.config";
 
+import 'rxjs/add/operator/catch';
+
 @Injectable()
 export class DefaultOAuthInterceptor implements HttpInterceptor {
-    
+
     constructor(
         private authStorage: OAuthStorage,
         private errorHandler: OAuthResourceServerErrorHandler,
@@ -24,7 +25,7 @@ export class DefaultOAuthInterceptor implements HttpInterceptor {
     }
 
     public intercept(req: HttpRequest<any>, next: HttpHandler): Observable<HttpEvent<any>> {
-        
+
         let url = req.url.toLowerCase();
 
         if (!this.moduleConfig) return next.handle(req);
@@ -33,22 +34,25 @@ export class DefaultOAuthInterceptor implements HttpInterceptor {
         if (!this.checkUrl(url)) return next.handle(req);
 
         let sendAccessToken = this.moduleConfig.resourceServer.sendAccessToken;
-        
-        if (sendAccessToken) {
+
+        if (sendAccessToken && this.authStorage.getItem('access_token')) {
 
             let token = this.authStorage.getItem('access_token');
             let header = 'Bearer ' + token;
 
             let headers = req.headers
-                                .set('Authorization', header);
+                .set('Authorization', header);
 
             req = req.clone({ headers });
         }
 
+<<<<<<< HEAD
+        return next.handle(req).catch(err => this.errorHandler.handleError(err));
+
+=======
         return next.handle(req).pipe(
             catchError(err => this.errorHandler.handleError(err))
         );
-
+>>>>>>> f0791bd3be17128862b73004a308a2c21488626c
     }
-
 }
