@@ -1,15 +1,12 @@
 import {HttpResponse} from '@angular/common/http';
-import { Observable,  } from 'rxjs/Observable';
-import {  _throw } from 'rxjs/observable/throw';
+import { Observable, throwError } from 'rxjs';
 
 export abstract class OAuthResourceServerErrorHandler {
     abstract handleError(err: HttpResponse<any>): Observable<any>;
 }
 
 export class OAuthNoopResourceServerErrorHandler implements OAuthResourceServerErrorHandler {
-    
     handleError(err: HttpResponse<any>): Observable<any> {
-        return _throw(err);
+        return throwError(err);
     }
-
 }
