@@ -5,11 +5,11 @@ export abstract class OAuthModuleConfig {
 }
 
 export abstract class OAuthResourceServerConfig {
-    
+
     /**
      * Urls for which calls should be intercepted.
      * If there is an ResourceServerErrorHandler registered, it is used for them.
-     * If sendAccessToken is set to true, the access_token is send to them too. 
+     * If sendAccessToken is set to true, the access_token is send to them too.
      */
     allowedUrls: Array<string>;
     sendAccessToken: boolean;
